% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ships.methods.R
\name{ais.to.DeponsShips}
\alias{ais.to.DeponsShips}
\title{Convert ship tracks to DeponsShips object}
\usage{
ais.to.DeponsShips(data, landsc, title = "NA")
}
\arguments{
\item{data}{data.frame with ship positions and the times at which the
positions were recorded}

\item{landsc}{A \code{DeponsRaster} object corresponding to the
landscape that the ships move in. It is assumed that the spatial projection
of the ship positions corresponds to that of the DeponsRaster object}

\item{title}{Title of the output object}
}
\value{
Returns a \code{DeponsShips} object
}
\description{
Crop one or more ship tracks to the extent of a landscape
and convert the data into a \code{DeponsShips-class} object. The in the
'pause' in the \code{DeponsShips} object corresponds to the number of half-hour
intervals where ships do not move, e.g. when in a port.
}
\examples{
data(aisdata)
plot(aisdata$x, aisdata$y, type="n", asp=1)
ids <- sort(unique(aisdata$id))
my.colors <- rainbow(length(ids))
for (i in length(ids)) {
  id <- ids[i]
  points(aisdata$x[aisdata$id==id], aisdata$y[aisdata$id==id],
     cex=0.6, col=my.colors[i])
}
data(bathymetry)
plot(bathymetry, add=TRUE)
depons.ais <- ais.to.DeponsShips(aisdata, bathymetry)
the.routes <- routes(depons.ais)
for (i in 1:length(ids)) {
  points(the.routes[[i]]$x, the.routes[[i]]$y,
         cex=0.6, pch=16, col=my.colors[i])
}
}
\seealso{
\code{\link{aisdata}} can be used as input to as.DeponsShips
}
