\docType{data}
\name{aisdata}
\alias{aisdata}
\title{Position for three ships in the inner Danish waters}
\format{
data.frame
}
\description{
Automatic identification system (AIS) data for three ships
in Kattegat and the Western Baltic from 20 Dec 2015. The data set includes
the variables id (the Maritime Mobile Service Identity number), time,
speed (in knots), type, length (in metres), x and y (which provide the
coordinates of the ship at a given time. The coordinates use the UTM zone 32
projection (CRS = "+proj=utm +zone=32 +units=m +no_defs +datum=WGS84").
}
\keyword{datasets}
