\name{argos.metrics}
\alias{argos.metrics}
\docType{data}
\title{Large-scale metrics of real animals}
\description{
This dataset summarizes large-scale movement metrics from 57 animals tracked 
in the North Sea and Skaggerak. For each individual, it includes home range size (km²), 
maximum Net Squared Displacement (km²), and a sinuosity index based on Benhamou (2004), 
which quantifies the tortuosity of movement paths. These metrics are intended for 
comparison with metrics of simulated tracks generated using the \code{calib_02} function.
}
\format{data.frame}
}
\keyword{datasets}
