% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{argosmetrics}
\alias{argosmetrics}
\title{Fine and Large-scale metrics of Argos data}
\format{
A nested list of data frames.
}
\description{
A nested list containing fine- and large-scale movement metrics
derived from satellite-tracked harbour porpoises (Argos data) in the North Sea
and Kattegat.

Fine-scale movement metrics include home range (HR, km2), mean net squared
displacement (NSD, km2), and residence time (RT, days), calculated from
30-day track segments during which animals used only fine-scale movements.

Large-scale movement metrics include HR (km2), maximum NSD (km2), a sinuosity
index (Benhamou, 2004), and cumulative distance moved (km), calculated from
100-day track segments irrespective of movement type. Large-scale metrics are
available for the North Sea only.
}
\keyword{datasets}
