% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn-methods.R
\name{as.data.frame,DeponsDyn-method}
\alias{as.data.frame,DeponsDyn-method}
\title{Convert DeponsDyn object to data frame}
\usage{
\S4method{as.data.frame}{DeponsDyn}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{DeponsDyn object}

\item{row.names}{NULL or a character vector giving the row names for the data
frame. Missing values are not allowed.}

\item{optional}{Logical (not used)}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
\code{data.frame} object
}
\description{
Function  for converting DEPONS population dynamics object
to a data frame.
}
\examples{
data(porpoisedyn)
class(porpoisedyn)
the.dyn <- as.data.frame(porpoisedyn)
}
