% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-methods.R
\name{as.data.frame,DeponsTrack-method}
\alias{as.data.frame,DeponsTrack-method}
\title{Convert DEPONS track to data frame}
\usage{
\S4method{as.data.frame}{DeponsTrack}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{DeponsTrack object}

\item{row.names}{NULL or a character vector giving the row names for the data
frame. Missing values are not allowed.}

\item{optional}{Logical (not used)}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
\code{data.frame} object
}
\description{
Function  for converting DEPONS movement track file to a
data frame.
}
\examples{
data(porpoisetrack)
class(porpoisetrack)
the.track <- as.data.frame(porpoisetrack)
}
