% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-methods.R, R/track-methods.R
\name{bbox}
\alias{bbox}
\alias{bbox,DeponsRaster-method}
\alias{bbox,DeponsTrack-method}
\title{Get bbox from DeponsRaster object}
\usage{
\S4method{bbox}{DeponsRaster}(obj)

\S4method{bbox}{DeponsTrack}(obj)
}
\arguments{
\item{obj}{DeponsTrack object}
}
\description{
Retrieves spatial bounding box from object.

Retrieves spatial bounding box from object.
}
