% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_methods.R
\name{calib_01}
\alias{calib_01}
\title{Plot distribution of turning angles, step lengths and speed of tracked simulated porpoises.}
\usage{
calib_01(depons_track)
}
\arguments{
\item{depons_track}{Object of class `DeponsTrack` produced by either read.DeponsTrack or read.DeponsTrackBatch}
}
\value{
Plot histograms of turning angles, step length and speed.
It also stores those metrics of the animal in a dataframe.
}
\description{
Plot distribution of turning angles, step lengths and speed of tracked simulated porpoises.
}
