% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_methods.R
\name{calib_03_make_sim_dpm}
\alias{calib_03_make_sim_dpm}
\title{Create piling-response table from simulated porpoise counts}
\usage{
calib_03_make_sim_dpm(
  sim.data,
  cpod.data,
  cpod.coord,
  piling.data,
  min.dists = c(1500, 3000, 6000, 9000, 12000, 15000),
  max.dists = c(3000, 6000, 9000, 12000, 15000, 18000),
  c.val,
  RT.val,
  tz,
  bad.data
)
}
\arguments{
\item{sim.data}{Data frame with at least columns `block`, `datetime`, `PorpoiseCount`.}

\item{cpod.data}{A data frame of observed CPOD data containing at least
the columns `station`, `time`, and `pod`. For each station, this table is
used to identify which CPOD was active and when pod switches occurred, so
that simulated records can be matched to the correct pod before aggregation.}

\item{cpod.coord}{CPOD coordinates table. Should contain at least columns `station`,
`x`, `y`. Station numbers should correspond to block numbers in the "block.asc" file
used to count porpoises per block in DEPONS.}

\item{piling.data}{Piling data table. Should contain at least columns `id`,
`x`, `y`, `start.time`,`stop.time`. The `start.time` and `stop.time` columns should be
date-time objects or character strings convertible to date-time format,
expressed in the same time zone as `sim.dat$datetime`.}

\item{min.dists}{Numeric vector of lower distance bounds in meters. This is to create
an interval of distance from piling, and dpm is aggregated per distance interval.}

\item{max.dists}{Numeric vector of upper distance bounds in meters.}

\item{c.val}{Simulation parameter value for `c`.}

\item{RT.val}{Simulation parameter value for `RT`.}

\item{tz}{Time zone.}

\item{bad.data}{A data frame with columns `date.start`, `date.end` (Date
objects), and `station` (numeric). One row per exclusion. Use `NULL`
(default) to treat all data as good.}
}
\value{
A data frame with aggregated DPM-vs-hour, with the same format as
observed DPM data.
}
\description{
Converts a simulation summary table into the aggregated
DPM-vs-hours-since-piling table used for response to piling calibration.
The function matches simulated porpoise counts to CPOD stations and pods,
assigns each record to the nearest piling event (found by the helper function
`find.nearest.piling`), groups records into distance intervals from piling (
using the helper function `get.pods.at.dist.sim.dat`), and aggregates detections
by piling event and hour since piling (using the helper function `aggregate.dpm.sim.dat`).
The output has the same general structure as the observed DPM data used for
comparison with simulations.
The default distance interval classes are 1.5-3 km, 3-6 km, 6-9 km, 9-12 km,
12-15 km and 15-18 km (modify parameters min.dists and max.dists to get different
intervals).
}
\examples{
\dontrun{
load(gemini.obs.data)
bad.data <- data.frame(date.start = as.Date("2010-06-20"),
date.end = as.Date("2010-06-25"),
station = c(7, 8))

dpm.sim <- calib_03_make_sim_dpm(sim.data = sim.data,
cpod.data = gemini.obs.data$cpod.data,
cpod.coord = gemini.obs.data$cpod.coord,
piling.data = gemini.obs.data$piling.data,
c.val = gemini.sim.data$sim.data$c,
RT.val = gemini.sim.data$sim.data$RT,
tz = "Europe/Amsterdam",
bad.data = bad.data)
}
}
