% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_methods.R
\name{calib_03_ssd}
\alias{calib_03_ssd}
\title{Compute SSD for one response to piling simulation}
\usage{
calib_03_ssd(sim.dpm, obs.dpm, filter.good.data = TRUE, plot = FALSE)
}
\arguments{
\item{sim.dpm}{Data frame returned by `calib_03_make_dpm_simdat()`.}

\item{obs.dpm}{Observed DPM-vs-hour data.}

\item{filter.good.data}{Logical. If `TRUE`, rows flagged as
`good.data == FALSE` are excluded from both `sim.dpm` and `obs.dpm`
before computing the SSD. Defaulft is `TRUE`.}

\item{plot}{Logical. If `TRUE`, the function plots the observed and scaled simulated
response curves for each distance class}
}
\value{
A list with `sum.ssd` and per-distance SSD table.
}
\description{
Computes the sum of squared differences (SSD) between observed
and simulated porpoise detection responses to piling, separately for each
distance class and summed across classes.

Observed and simulated aggregated DPM data are first summarised by hour since
piling. Simulated values are then scaled to have the same overall mean and
standard deviation as the observed values before SSD is calculated.
}
\examples{
\dontrun{
load(gemini.obs.data)
load(gemini.sim.data)

calib_03_ssd(sim.dpm=gemini.sim.data$dpm.sim,
obs.dpm = gemini.obs.data$dpm.vs.hr, plot=T)
}
}
