\docType{data}
\name{coastline}
\alias{coastline}
\title{Coastline of Northern Europe}
\format{
SpatialPolygonsDataFrame
}
\description{
An object of class \code{\link[sp]{SpatialPolygonsDataFrame}} showing the
coastline of the North Sea, Kattegat, and the Western Baltic. The map projection
used is ETRS89 – EPSG:3035 projection as for the North Sea raster files used
by DEPONS. The corresponding proj4string is "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000
+y_0=3210000 +datum=WGS84 +units=m +no_defs".
}
\keyword{datasets}
