% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-methods.R, R/track-methods.R
\name{crs}
\alias{crs}
\alias{crs,DeponsRaster-method}
\alias{crs,DeponsTrack-method}
\title{Get map projection}
\usage{
\S4method{crs}{DeponsRaster}(x)

\S4method{crs}{DeponsTrack}(x)
}
\arguments{
\item{x}{Object of class {code{DeponsRaster}} or \code{DeponsTrack}.}
}
\description{
Get the map projection (also known as coordinate reference
system, crs) of DeponsRaster and DeponsTrack objects.
}
\seealso{
See \code{\link[sp]{proj4string}} about the format of the text
string defining the map projection.
}
