% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-misc.R
\name{DEPONS2R}
\alias{DEPONS2R}
\title{Package for analyzing data simulated using the DEPONS model}
\description{
Classes and methods for analyzing and plotting changes in
population size for different parts of the landscape (i.e. different 'blocks')
}
