% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockdyn-methods.R
\docType{class}
\name{DeponsBlockdyn-class}
\alias{DeponsBlockdyn-class}
\title{DeponsBlockdyn-class and methods}
\description{
Classes and methods for analyzing and plotting changes in
population size for different parts of the landscape (i.e. different 'blocks')
}
\details{
The data frame with simulation output includes the columns 'tick',
which indicates the number of half-hourly time steps since the start of the
simulation; a column 'count.X' with the number of animals in each block at
each tick, and 'real.time', which shows the real-world equivalent to 'tick.
}
\section{Slots}{

\describe{
\item{\code{title}}{Character. Name of the object or simulation}

\item{\code{landscape}}{Character. Identifier for the landscape used in the DEPONS
simulations. The landscapes 'DanTysk', 'Gemini', 'Kattegat', 'North Sea',
'Homogeneous', and 'User defined' are distributed with the DEPONS model.}

\item{\code{simtime}}{\code{\link{POSIXlt}} object with the date and time when the simulation was
finished. This is read from the name of the imput file.}

\item{\code{simstart}}{POSIXlt object with the first day of the simulation, i.e.
the first day in the period that the simulations are intended to represent in
the real world.}

\item{\code{data}}{Data frame with simulation output.}
}}

\note{
DeponsBlockdyn-objects are usually read in from csv files produced during
DEPONS simulations. These files are named 'PorpoisePerBlock.XXX.csv', where XXX
indicates the date and time when the simulation was finished.
}
\examples{
a.DeponsBlockdyn <- new("DeponsBlockdyn")
a.DeponsBlockdyn
}
\seealso{
\code{\link[DEPONS2R]{plot.DeponsBlockdyn}} and
\code{\link[DEPONS2R]{read.DeponsBlockdyn}}.
}
