% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-methods.R
\docType{class}
\name{DeponsTrack-class}
\alias{DeponsTrack-class}
\title{DeponsTrack-class and methods}
\description{
Classes and methods for manipulating and plotting movement
tracks generated with DEPONS.
}
\section{Slots}{

\describe{
\item{\code{title}}{Name of the object (character)}

\item{\code{simdate}}{POSIXlt object with the date and time when the simulation was
finished. This is read from the name of the imput file.}

\item{\code{crs}}{CRS object providing the coordinate reference system used; see
\code{\link[sp]{CRS}} for details}

\item{\code{tracks}}{Listwith one or more tracks, each stored
as a \code{\link[sp]{SpatialPointsDataFrame}} object)}
}}

\seealso{
\link{read.DeponsTrack}
}
