% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-accessors.R
\name{dyn}
\alias{dyn}
\alias{dyn,DeponsDyn-method}
\alias{dyn,DeponsBlockdyn-method}
\title{Extract the dynamics data from DeponsDyn objects}
\usage{
\S4method{dyn}{DeponsDyn}(x, value)

\S4method{dyn}{DeponsBlockdyn}(x, value)
}
\arguments{
\item{x}{Object of class \code{DeponsBlockdyn}.}

\item{value}{Character string}
}
\description{
Extract the dynamics data from DeponsDyn objects

Extract the dynamics dyn from DeponsDyn objects
}
