% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geminidata}
\alias{geminidata}
\title{Observed and simulated data from Gemini windfarm construction}
\format{
A nested list with two elements:
\describe{
  \item{obs.data}{A list containing:
    \describe{
      \item{piling.data}{Data frame of gemini piling coordinates and schedule.}
      \item{cpod.coord}{Data frame of CPOD station coordinates.}
      \item{cpod.data}{Observed CPOD detection data.}
      \item{dpm.vs.hr}{Observed DPM data aggregated by distance to piling and
      hour since piling.}
    }
  }
  \item{sim.data}{A list containing:
    \describe{
      \item{sim.data}{Raw simulated porpoise counts per block.}
      \item{sim.dpm.vs.hr}{Simulated DPM data aggregated by distance interval
      and hour since piling.}
    }
  }
}
}
\description{
A nested list containing observed and simulated data used for
response-to-piling calibration in the Gemini case study.

The observed data component contains CPOD detections, CPOD coordinates,
gemini piling coordinates and schedule, and observed DPM data aggregated by
distance to piling and hour since piling. `cpod.data`, `cpod.coord`, and
`piling.data` are used in [calib_03_make_sim_dpm()], and `obs.dpm` is used
to calculate SSD with the function [calib_03_ssd()].

The simulated data component contains raw simulated porpoise counts per block
(`sim.data`; summed across 10000 replicates) resulting from the best combination of
c and RT in DEPONS 4.0, and those simulated data aggregated by
distance interval and hour since piling (`sim.dpm`; aggregated with the function
[calib_03_make_sim_dpm()]). These simulated DPM data can be visualised against
observed ones with the function [calib_03_ssd()].
}
\keyword{datasets}
