% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-misc.R
\name{get.latest.sim}
\alias{get.latest.sim}
\title{Get output}
\usage{
get.latest.sim(dir)
}
\arguments{
\item{dir}{Directory to look for simulation output in (character string)}
}
\description{
Returns the name of the newest simulation output within the
specified directory
}
