% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-accessors.R
\name{landscape<-}
\alias{landscape<-}
\alias{landscape}
\alias{landscape,DeponsTrack-method}
\alias{landscape<-,DeponsTrack-method}
\alias{landscape,DeponsDyn-method}
\alias{landscape<-,DeponsDyn-method}
\title{Get or set the landscape name of DeponsTrack objects}
\usage{
\S4method{landscape}{DeponsTrack}(x) <- value

\S4method{landscape}{DeponsTrack}(x, value)

\S4method{landscape}{DeponsDyn}(x) <- value

\S4method{landscape}{DeponsDyn}(x, value)
}
\arguments{
\item{x}{Object of class \code{DeponsDyn}.}

\item{value}{Character string}
}
\description{
Get or set the landscape name of DeponsTrack objects

Get or set the landscape of DeponsDyn objects
}
