% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-methods.R
\name{make.blocksraster}
\alias{make.blocksraster}
\alias{make.blocksraster,DeponsRaster-method}
\title{Create a new blocks file}
\usage{
\S4method{make.blocksraster}{DeponsRaster}(
  template,
  blocks,
  blockvals,
  fname,
  plot = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{template}{DeponsRaster object used as template for new blocks file}

\item{blocks}{List of areas to be used for new blocks. Each item in 'blocks' should
either be a matrix (with two columns, corresponding to x and y-coordinates)
or a \code{\link{SpatialPolygons}} object.}

\item{blockvals}{Vector of integer values defining the labels of the new blocks.
The first value defines the background value, so the length of 'blockvals'
should equal the number of blocks plus 1.}

\item{fname}{Name of the output file (ending with '.asc'). The blocks raster
will only be saved if a name is provided.}

\item{plot}{Whether a plot should be produced.}

\item{overwrite}{Overwrite existing file?}
}
\description{
Make file for counting the number of animals in different parts
('blocks') of the landscape during DEPONS simulations.
}
\details{
Any of the raster files in simulations for a specific landscape can
be used as template. The new blocks file will have the same resolution and
extent as the template, but will consist of a number of blocks with a shape
defined using the \code{blocks} parameter. If an item on the blocks list is
a matrix with spatial coordinates, it is assumed to use the same coordinate
reference system as the template. In that case the new block is defined by the
smallest possible rectangle that contains all positions. For \code{SpatialPoints}
or \code{SpatialPolygons} elements, the blocks can take any shape. Raster cells
are included in the new blocks if the polygons that define them include the
centre of the raster cell.
}
\note{
Blocks files should not be modified when running simulations in the
Kattegat landscape. In this landscape the blocks file allows simulated animals
to navigate beteween different parts of the landscape when they use large-scale
movements.
}
\examples{
# Make list with areas to designate as blocks
new.blocks <- list()
x <- runif(7, 700000, 760000)
y <- runif(7, 6200000, 6300000)
new.blocks[[1]] <- cbind(x,y)
x <- c(600000, 635000, 670000, 635000)
y <- c(6150000, 6200000, 6150000, 6100000)
xy <- cbind(x,y)
library(sp)
srl <- list(Polygon(xy))
Srl <- list(sp::Polygons(srl, ID=as.vector(as.character(2))))
new.blocks[[2]] <- SpatialPolygons(Srl, proj4string=crs(bathymetry))

make.blocksraster(template=bathymetry, blocks=new.blocks, plot=TRUE)
points(new.blocks[[1]])
plot(new.blocks[[2]], add=TRUE)

\dontrun{
make.blocksraster(template=bathymetry, blocks=new.blocks, fname="test.asc")
}
}
\seealso{
See \code{readShapePoly} (\code{maptools}-package) for details on how to
read a polygon shapefile into a SpatialPolygonsDataFrame object.
}
