% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-misc.R
\name{make.windfarms}
\alias{make.windfarms}
\title{Make wind farm scenario}
\usage{
make.windfarms(area.file, area.def, n.wf, n.turb, turb.dist, min.wf.dist)
}
\arguments{
\item{area.file}{Name of the raster file specifying where the wind farms
should be constructed.}

\item{area.def}{Value in \code{area.file} for the areas were wind farms can
be located}

\item{n.wf}{Number of wind farms to construct}

\item{n.turb}{Total number of turbines to construct}

\item{turb.dist}{Distance between turbines within a wind farm (meters)}

\item{min.wf.dist}{Minimum distance between wind farms (meters)}
}
\description{
Produce a hypothetical wind farm construction scenario, specifying
the position and timing of individual piling events, as well as the sound
source level. All wind farms are assumed to consist of the same number of
turbines, laid out in a rectangular grid.
}
