% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn-methods.R
\name{plot,DeponsDyn,missing-method}
\alias{plot,DeponsDyn,missing-method}
\alias{plot.DeponsDyn}
\title{Plot a DeponsDyn object}
\usage{
\S4method{plot}{DeponsDyn,missing}(x, y, ...)
}
\arguments{
\item{x}{DeponsDyn object}

\item{y}{Not used}

\item{...}{Optional plotting parameters}
}
\description{
Plot population dynamics simulated with DEPONS
}
\examples{
data("porpoisedyn")
plot(porpoisedyn)
}
