% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ships.methods.R
\name{plot,DeponsShips,missing-method}
\alias{plot,DeponsShips,missing-method}
\alias{plot.DeponsShips}
\title{Plot a DeponsShips object}
\usage{
\S4method{plot}{DeponsShips,missing}(x, y, ...)
}
\arguments{
\item{x}{DeponsShips object}

\item{y}{Not used}

\item{...}{Optional plotting parameters}
}
\description{
Plot the tracks that ship agents move along in DEPONS.
}
