% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_methods.R
\name{plot_calib02}
\alias{plot_calib02}
\title{Plotting of simulated fine or large-scale metrics against argos data}
\usage{
plot_calib02(sim.metrics, option)
}
\arguments{
\item{sim.metrics}{A dataframe of movement metrics obtained with the calib_02 function}

\item{option}{A character string, either `"fine"` or `"large"`, indicating which set of Argos metrics (fine-scale or large-scale)
to plot against simulated ones.}
}
\value{
a plot of real (Argos) vs simulated metrics
}
\description{
Plotting of simulated fine or large-scale metrics against argos data
}
