% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{porpoisebdyn}
\alias{porpoisebdyn}
\title{Simulated porpoise population dynamics}
\format{
DeponsBlockdyn
}
\description{
An object of class \code{DeponsBlockdyn} with output from a DEPONS
simulation based on the North Sea landscape, using a landscape divided into
two blocks.
}
\seealso{
\code{\link{DeponsBlockdyn-class}}
}
\keyword{datasets}
