% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{porpoisetrack}
\alias{porpoisetrack}
\title{Simulated porpoise track}
\format{
DeponsTrack
}
\description{
An object with three elements: \code{title}, \code{date}, and
\code{tracks}, where \code{tracks} is a list of objects of class
\code{\link[sp]{SpatialPointsDataFrame}}, each of which corresponds to one
simulated animal. \code{date} is the simulation date
}
\keyword{datasets}
