% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockdyn-methods.R
\name{read.DeponsBlockdyn}
\alias{read.DeponsBlockdyn}
\title{Reading simulated population count for blocks}
\usage{
read.DeponsBlockdyn(
  fname,
  title = "NA",
  landscape = "NA",
  simtime = "NA",
  simstart = "NA"
)
}
\arguments{
\item{fname}{Name of the file (character) that contains movement data
generated by DEPONS. The name includes the path to the directory if this is
not the current working directory.}

\item{title}{Optional character string giving name of simulation}

\item{landscape}{The landscape used in the simulation}

\item{simtime}{Optional POSIXlt object with date of simulation. If
not provided this is obtained from name of input file}

\item{simstart}{The start of the period that the  simulation represents, i.e.
the real-world equivalent of 'tick 1' (POSIXlt)}
}
\description{
Function for reading DEPONS simulation output with number of
animals per block for each time step.
}
\seealso{
See \code{\link{DeponsBlockdyn-class}} for details on what is stored in
the output object.
}
