% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn-methods.R
\name{read.DeponsDyn}
\alias{read.DeponsDyn}
\title{Reading DEPONS simulation output}
\usage{
read.DeponsDyn(
  fname,
  title = "NA",
  landscape = "NA",
  simtime = "NA",
  startday = "NA"
)
}
\arguments{
\item{fname}{Name of the file (character) that contains movement data
generated by DEPONS. The name includes the path to the directory if this is
not the current working directory.}

\item{title}{Optional character string giving name of simulation}

\item{landscape}{The landscape used in the simulation}

\item{simtime}{Optional POSIXlt object with the date and time when the
simulation finished. If not provided this is obtained from name of input file}

\item{startday}{The start of the period that the  simulation represents, i.e.
the real-world equivalent of 'tick 1' (POSIXlt)}
}
\description{
Function  for reading simulation output produced by DEPONS.
}
\seealso{
See \code{\link{DeponsDyn-class}} for details on what is stored in
the output object.
}
