% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-accessors.R
\name{startday}
\alias{startday}
\alias{startday,DeponsBlockdyn-method}
\alias{startday,DeponsDyn-method}
\alias{startday<-}
\alias{startday<-,DeponsBlockdyn-method}
\alias{startday<-,DeponsDyn-method}
\title{Get or set start date for simulation}
\usage{
\S4method{startday}{DeponsBlockdyn}(x)

\S4method{startday}{DeponsDyn}(x)

\S4method{startday}{DeponsBlockdyn}(x) <- value

\S4method{startday}{DeponsDyn}(x) <- value
}
\arguments{
\item{x}{Object of class \code{DeponsDyn}}

\item{value}{POSIXlt or character string of the form 'yyyy-mm-dd'}
}
\description{
Get or set start date for simulation

Get or set start date for simulation
}
\details{
The start date indicates the start of the period that the simulation
is supposed to represent.

The start date indicates the start of the period that the simulation
is supposed to represent.
}
\note{
The assignment of a new start time is currently quite time consuming.
}
