% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockdyn-methods.R, R/dyn-methods.R,
%   R/raster-methods.R, R/ships.methods.R, R/track-methods.R
\name{Summary-methods}
\alias{Summary-methods}
\alias{summary}
\alias{summary,DeponsBlockdyn-method}
\alias{summary,DeponsDyn-method}
\alias{summary,DeponsRaster-method}
\alias{summary,DeponsShips-method}
\alias{summary,DeponsTrack-method}
\title{Summary}
\usage{
\S4method{summary}{DeponsBlockdyn}(object)

\S4method{summary}{DeponsDyn}(object)

\S4method{summary}{DeponsRaster}(object)

\S4method{summary}{DeponsShips}(object)

\S4method{summary}{DeponsTrack}(object)
}
\arguments{
\item{object}{Depons* object}
}
\value{
list summarizing the DeponsBlockdyn object

table summarizing the DeponsBlockdyn object

list summarizing the DeponsRaster object

list summarizing the DeponsTrack object
}
\description{
Summarizes different kinds of objects created based on output
  from the DEPONS model
}
\details{
The summary method is available for \code{\link{DeponsTrack-class}},
\code{\link{DeponsDyn-class}}, \code{\link{DeponsRaster-class}},
and \code{\link{DeponsBlockdyn-class}}-objects.
}
