% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-misc.R
\name{tick.to.date}
\alias{tick.to.date}
\title{Convert tick number to date}
\usage{
tick.to.date(tick, ...)
}
\arguments{
\item{tick}{Numeric; tick number}

\item{...}{Other parameters used by as.POSIXlt}
}
\description{
Converts the number of ticks since the start of the simulation
to a specific data while taking into account that DEPONS assumes that there
is 30 days per month
}
