% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn-methods.R, R/track-methods.R
\name{title<-}
\alias{title<-}
\alias{title}
\alias{title,DeponsTrack-method}
\alias{title<-,DeponsTrack-method}
\title{Get or set the title of DeponsTrack objects}
\usage{
title(x, value)

\S4method{title}{DeponsTrack}(x) <- value

\S4method{title}{DeponsTrack}(x, value)
}
\arguments{
\item{x}{Object of class \code{DeponsTrack}.}

\item{value}{Character string}
}
\description{
Get or set the title of DeponsTrack objects
}
