\name{argosmetrics}
\alias{argosmetrics}
\docType{data}
\title{Fine and Large-scale metrics of Argos data}
\description{
This dataset is a list of two dataframes containing fine and large-scale movements metrics of satellite tracked animals (Argos). 
For fine-scale movements, metrics include home range (HR, km2), mean net squared displacement (NSD, km2) and 
residence time (Rt, days) on 30-days tracks where animals used only fine-scale movements. For large-scale movements, 
metrics include HR (km2), max NSD (km2), a sinuosity index (Benhamou, 2004) and cumulative distance moved (km) on 100-days tracks that are independent from movement types.}
\format{list}
}
\keyword{datasets}
