% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-methods.R, R/track-methods.R
\name{bbox}
\alias{bbox}
\alias{bbox,DeponsRaster-method}
\alias{bbox,DeponsTrack-method}
\title{Get bbox from Depons* object}
\usage{
\S4method{bbox}{DeponsRaster}(obj)

\S4method{bbox}{DeponsTrack}(obj)
}
\arguments{
\item{obj}{DeponsRaster or DeponsTrack object}
}
\value{
Returns a \code{matrix} defining the northern, southern, eastern and
western boundary of a DeponsRaster object or of one or more DeponsTrack
objects.
}
\description{
Retrieves spatial bounding box from object. If a Depons* object
is a DeponsTrack object containing multiple track, the box bounds all tracks.
}
\seealso{
\code{\link{make.clip.poly}}
}
