% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-methods.R, R/z-accessors.R
\name{crs}
\alias{crs}
\alias{crs,DeponsRaster-method}
\alias{crs,DeponsTrack-method}
\alias{crs<-,DeponsTrack-method}
\alias{crs<-}
\title{Get or set map projection}
\usage{
\S4method{crs}{DeponsRaster}(x)

\S4method{crs}{DeponsTrack}(x)

\S4method{crs}{DeponsTrack}(x) <- value
}
\arguments{
\item{x}{Object of class class {code{DeponsRaster}} or \code{DeponsTrack}}

\item{value}{\code{\link{proj4string}} identifying the map projection}
}
\description{
Get or set the map projection (also known as coordinate reference
system, crs) of DeponsRaster and DeponsTrack objects. For {sp} objects the
text string defining the crs is called the \code{\link[sp]{proj4string}}.
}
