% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-accessors.R
\name{crs}
\alias{crs}
\alias{crs,DeponsTrack-method}
\alias{crs<-,DeponsTrack-method}
\alias{crs<-,DeponsShips-method}
\alias{crs,DeponsRaster-method}
\alias{crs,DeponsShips-method}
\alias{crs<-}
\alias{crs<-,DeponsRaster-method}
\title{Get or set map projection in Depons* objects}
\usage{
\S4method{crs}{DeponsTrack}(x)

\S4method{crs}{DeponsShips}(x)

\S4method{crs}{DeponsRaster}(x)

\S4method{crs}{DeponsTrack}(x) <- value

\S4method{crs}{DeponsShips}(x) <- value

\S4method{crs}{DeponsRaster}(x) <- value
}
\arguments{
\item{x}{Object of class class \code{DeponsRaster}, \code{DeponsShips} or
\code{DeponsTrack}}

\item{value}{(proj4string) identifying the map projection}
}
\description{
Get or set the map projection (also known as coordinate reference
system, crs) of DeponsRaster and DeponsTrack objects.
}
