% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-misc.R
\name{DEPONS2R}
\alias{DEPONS2R}
\title{Package for analyzing DEPONS simulation output}
\description{
Methods for analyzing population dynamics and movement tracks
simulated using the DEPONS model (v.3.0; \url{https://www.depons.eu}), for
manipulating input raster files, shipping routes and for analyzing sound
propagated from ships.

The classes used in DEPONS2R include:
\itemize{
\item \code{\linkS4class{DeponsTrack}} movement tracks, read from "RandomPorpoise.XXX.csv" files
\item \code{\linkS4class{DeponsDyn}} population dynamics data, from "Statistics.XXX.csv" files
\item \code{\linkS4class{DeponsBlockdyn}} data from "PorpoisePerBlock.XXX.csv" files
\item \code{\linkS4class{DeponsShips}} data from "ships.json" files or from AIS data
}
Here the \code{DeponsDyn} data include both changes in population size and energetics
through time for the entire landscape, whereas \code{DeponsBlockdyn} data include
variations in population size in different parts (or 'blocks') of the landscape.
XXX is the date and time when the simulation was finished.
}
