% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-misc.R
\name{DEPONS2R}
\alias{DEPONS2R}
\title{Package for analyzing DEPONS simulation output}
\description{
Classes and methods for analyzing and plotting movement tracks
and population dynamics simulated using the DEPONS model (\url{http://www.depons.eu}).

The types of simulated data include:
\itemize{
\item \code{\linkS4class{DeponsTrack}} movement tracks, read from "RandomPorpoise.XXX.csv" files
\item \code{\linkS4class{DeponsDyn}} population dynamics data, from "Statistics.XXX.csv" files
\item \code{\linkS4class{DeponsBlockdyn}} data from "PorpoisePerBlock.XXX.csv" files
}
Here the \code{DeponsDyn} data include both changes in population size and energetics
through time for the entire landscape, whereas \code{DeponsBlockdyn} data include
variations in population size in different parts (or 'blocks') of the landscape.
XXX is the date and time when the simulation was finished.
}
