% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-methods.R
\docType{class}
\name{DeponsTrack-class}
\alias{DeponsTrack-class}
\title{DeponsTrack-class}
\description{
Stores objects containing animal movement tracks simulated using
the DEPONS model

Classes for manipulating and plotting movement
tracks generated with DEPONS.
}
\section{Slots}{

\describe{
\item{\code{title}}{Name of the object (character)}

\item{\code{landscape}}{Name of the object (character)}

\item{\code{simtime}}{POSIXlt object with the date and time when the simulation was
finished. This is read from the name of the imput file.}

\item{\code{crs}}{CRS object providing the coordinate reference system used; see
\code{\link[sp]{CRS}} for details}

\item{\code{tracks}}{Listwith one or more tracks, each stored
as a \code{\link[sp]{SpatialPointsDataFrame}} object)}
}}

\seealso{
\code{\link[DEPONS2R]{plot.DeponsTrack}} and
\code{\link[DEPONS2R]{read.DeponsTrack}}
}
