% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-accessors.R
\name{dyn}
\alias{dyn}
\alias{dyn,DeponsDyn-method}
\alias{dyn,DeponsBlockdyn-method}
\title{Extract population dynamics from objects}
\usage{
\S4method{dyn}{DeponsDyn}(x)

\S4method{dyn}{DeponsBlockdyn}(x)
}
\arguments{
\item{x}{Object of class \code{DeponsBlockdyn}.}
}
\description{
Extract population dynamics from objects
}
