% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-misc.R
\name{get.latest.sim}
\alias{get.latest.sim}
\title{Get name of newest file}
\usage{
get.latest.sim(type = "dyn", dir)
}
\arguments{
\item{type}{Type of simulation output to check; can be one of:
"dyn" (for looking in "Statistics.XX.csv" files),
"blockdyn" (for looking in "PorpoisePerBlock.XX.csv" files)
"track" (for looking in "RandomPorpoise.XX.csv" files)}

\item{dir}{Directory to look for simulation output in (character string)}
}
\description{
Returns the name of the newest simulation output of a particular
type within the specified directory. The date and time are extracted from the
file name.
}
\seealso{
\code{\link{read.DeponsBlockdyn}} for example.
}
