% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-misc.R
\name{get.simtime}
\alias{get.simtime}
\title{Get simulation date}
\usage{
get.simtime(fname = NULL)
}
\arguments{
\item{fname}{Character string with name of the file to extract the simulation
date from, including the path}
}
\description{
Returns the date and time when a specific simulation was finished
(a \code{\link{POSIXlt}} object)
}
\seealso{
\code{\link{get.latest.sim}}
}
