% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ships.methods.R
\name{plot,DeponsShips,missing-method}
\alias{plot,DeponsShips,missing-method}
\alias{plot.DeponsShips}
\title{Plot a DeponsShips object}
\usage{
\S4method{plot}{DeponsShips,missing}(x, y, ...)
}
\arguments{
\item{x}{DeponsShips object}

\item{y}{Not used}

\item{...}{Optional plotting parameters, including 'col', 'main',
'add.legend', and 'legend.xy' (defaults to 'topright' when add.legend=TRUE)}
}
\description{
Plot the tracks that ship agents move along in DEPONS.
}
\examples{
data(ships)
plot(ships, col=c("red", "green", "blue"))

# convert route coordinate units from 'grid squares' to UTM
data(bathymetry)
out <- summary(bathymetry)
left <- out[[4]][1]
bottom <- out[[4]][2]
for (i in 1:3) {
    newroute <- ships@routes[[2]][[i]]*400
    newroute$x <- newroute$x + as.numeric(left)
    newroute$y <- newroute$y + as.numeric(bottom)
    ships@routes[[2]][[i]] <- newroute
    }

# Reproject coastline and clip to size of Kattegat landscape
library(sp)
data(bathymetry)
data(coastline)
coastline2 <- spTransform(coastline, crs(bathymetry))
bbox <- bbox(bathymetry)
clip.poly <- make.clip.poly(bbox, crs(bathymetry))
new.coastline <- rgeos::gIntersection(coastline2, clip.poly, byid = TRUE,
  drop_lower_td = TRUE)
\donttest{
plot(new.coastline, col="lightyellow2")
plot(ships, col=c("red", "green", "blue"), add=TRUE, add.legend=TRUE)
plot(clip.poly, add=TRUE)
}
}
