\docType{data}
\name{porpoisebdyn}
\alias{porpoisebdyn}
\title{Simulated porpoise population dynamics}
\format{
DeponsBlockdyn
}
\description{
An object of class \code{DeponsBlockdyn} with output from a DEPONS
simulation based on the North Sea landscape, using a landscape divided into
two blocks. Numbers of animals are counted per block.
}
\seealso{
\code{\link{DeponsBlockdyn-class}}, \code{\link{porpoisedyn}}
}
\keyword{datasets}
