\docType{data}
\name{porpoisedyn}
\alias{porpoisedyn}
\title{Simulated porpoise population dynamics}
\format{
DeponsDyn
}
\description{
An object of class \code{DeponsDyn} with output from a DEPONS
simulation based on the Kattegat landscape, assuming that the simulation
represents the period 2010-01-01 onward in the real world. Number of animals
and energy availability is recorded for the entire landscape.
}
\seealso{
\code{\link{DeponsDyn-class}}, \code{\link{porpoisebdyn}}
}
\keyword{datasets}
