% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{porpoisetrack}
\alias{porpoisetrack}
\title{Simulated porpoise track}
\format{
DeponsTrack
}
\description{
An object with four elements: \code{title}, \code{simdate}, \code{crs},
 and \code{tracks}. The \code{tracks} elemet is a list of objects of class
\code{\link[sp]{SpatialPointsDataFrame}}, each of which corresponds to one
simulated animal. \code{simdate} is the simulation date.
}
\keyword{#'}
\keyword{@seealso}
\keyword{\code{\link{DeponsTrack-class}}}
\keyword{datasets}
