% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ships.methods.R
\name{routes}
\alias{routes}
\alias{routes,DeponsShips-method}
\alias{routes<-,DeponsShips-method}
\alias{routes<-}
\title{Get or define routes in DeponsShips objects}
\usage{
\S4method{routes}{DeponsShips}(x)

\S4method{routes}{DeponsShips}(x) <- value
}
\arguments{
\item{x}{Object of class \code{DeponsShips}}

\item{value}{list with one named element per shipping route. Each element is
a data frame with the variables x and y, which define the coordinates of the fix-points
on the shipping routes.}
}
\description{
Get or define routes in DeponsShips objects
}
\seealso{
\code{\link{ships}}
}
