% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ships_methods.R
\docType{data}
\name{shipdata}
\alias{shipdata}
\title{Ships on example routes through the Kattegat}
\format{
DeponsShips object
}
\description{
The routes of fifteen ships of different types in the Kattegat during a period of 15 days. The fix points that define the routes use the UTM zone 32 projection (CRS = "+proj=utm +zone=32 +units=m +no_defs +datum=WGS84"; EPSG:32632; see https://epsg.io/32632).
}
\seealso{
[DeponsShips-class]
}
