% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ships.methods.R
\name{ships}
\alias{ships}
\alias{ships,DeponsShips-method}
\alias{ships<-,DeponsShips-method}
\alias{ships<-}
\title{Get or define ships in DeponsShips objects}
\usage{
\S4method{ships}{DeponsShips}(x)

ships(x) <- value
}
\arguments{
\item{x}{Object of class \code{DeponsShips}}

\item{value}{data frame with the 'name', 'speed', 'impact', and 'route' of
ships to be simulated. Here 'impact' is the sound source level (dB SPL) and
'route' is one of the shipping routes already defined in the DeponsShips object.}
}
\description{
Get or define ships in DeponsShips objects
}
\examples{
data(shipdata)
ships(shipdata)
}
\seealso{
\code{\link{routes}}
}
