% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ships_methods.R
\name{ships}
\alias{ships}
\alias{ships,DeponsShips-method}
\alias{ships<-,DeponsShips-method}
\alias{ships<-}
\title{Get or define ships in DeponsShips objects}
\usage{
\S4method{ships}{DeponsShips}(x)

ships(x) <- value
}
\arguments{
\item{x}{Object of class \code{DeponsShips}}

\item{value}{data frame with the 'name', 'type', 'length', and 'route' of
ships to be simulated, as well as 'tickStart' and 'tickEnd' defining when
the ships are to be included in simulations. 'route' is one of the shipping
routes defined in the DeponsShips object.}
}
\description{
Get or define ships in DeponsShips objects
}
\examples{
data(shipdata)
ships(shipdata)
}
\seealso{
\code{\link{routes}}
}
