% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-misc.R
\name{tick.to.time}
\alias{tick.to.time}
\title{Convert tick number to date}
\usage{
tick.to.time(tick, ...)
}
\arguments{
\item{tick}{Numeric; tick number}

\item{...}{Time zone (tz) and other parameters}
}
\description{
Converts the number of ticks since the start of the simulation
to a specific date while taking into account that DEPONS assumes that there
is 30 days per month. Returns an object of class \code{\link{as.POSIXlt}}
}
\note{
The conversion from ticks to date is problematic in February, with <30
days. Here time is forced to 'stands still'.
}
