% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-accessors.R
\name{title<-}
\alias{title<-}
\alias{title}
\alias{title,DeponsTrack-method}
\alias{title<-,DeponsTrack-method}
\alias{title<-,DeponsDyn-method}
\alias{title,DeponsDyn-method}
\title{Get or set the title of DeponsTrack objects}
\usage{
\S4method{title}{DeponsTrack}(x) <- value

\S4method{title}{DeponsTrack}(x, value)

\S4method{title}{DeponsDyn}(x) <- value

\S4method{title}{DeponsDyn}(x, value)
}
\arguments{
\item{x}{Object of class \code{DeponsDyn}.}

\item{value}{Character string}
}
\description{
Get or set the title of DeponsTrack objects

Get or set the title of DeponsDyn objects
}
