% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-accessors.R
\name{title<-}
\alias{title<-}
\alias{title<-,DeponsDyn-method}
\alias{title<-,DeponsShips-method}
\alias{title}
\alias{title,DeponsTrack-method}
\alias{title<-,DeponsTrack-method}
\alias{title,DeponsDyn-method}
\alias{title,DeponsShips-method}
\title{Get or set the title of Depons* objects}
\usage{
\S4method{title}{DeponsTrack}(x) <- value

\S4method{title}{DeponsDyn}(x) <- value

\S4method{title}{DeponsShips}(x) <- value

\S4method{title}{DeponsTrack}(x, value)

\S4method{title}{DeponsDyn}(x, value)

\S4method{title}{DeponsShips}(x, value)
}
\arguments{
\item{x}{Object of class \code{DeponsTrack}, \code{DeponsDyn},
\code{DeponsBlockdyn} or \code{DeponsShips}}

\item{value}{Character string}
}
\description{
Get or set the title of Depons* objects
}
